somehow it's working (maybe)

+++ lib/audio/audio.h
@@ -7,7 +7,6 @@
 #include <string.h>
 #include <unistd.h>
 #include <assert.h>
-#include <complex.h>
 
 // WAV header spec information:
 // https://docs.fileformat.com/audio/wav/
@@ -31,7 +30,7 @@ typedef struct wav_header {
 
 typedef struct wav_data_chunk {
     int32_t size;
-    int32_t *channel_data;
+    double *channel_data;
 } wav_data_chunk;
 
 void wav_read_audio_file(const char *filename, wav_header **header, wav_data_chunk **data_chunk);
